requirements.txt
Flask~=1.1.2
datetime~=4.3
uuid~=1.30
Flask-SQLAlchemy
PyJWT~=1.7.1
Werkzeug~=1.0.1